backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
PyJWT==2.8.0
redis==5.0.1
email-validator==2.1.0